Split Git auth and staging into its own steps
I am trying to figure out how to make the bail step skip the remaining steps but still have the pipeline to succeed if there is nothing to commit. I will probably need to use step outputs for this somehow. But not sure how yet, so I am leaving the rest in one step and will figure it out and break it apart later.