Repeat diff-index for debugging and use step output for condition
I am confused right now because in the last run the output was printed as "false" in the echo meaning we should not bail because we have changes, but then the commit command failed later, so let's see if we indeed have changes. Also I changed the Bash if syntax back just in case the semicolon was doing something.