Rename the workflow to just create-react-app
I have updated both the name of the workflow file and the name in the file. I want the file name to match what the workflow does and the formatting of the actual workflow name I think is better off without the verb and the space so it looks better and more regular to what these names usually look like on the badge.